Change ~ don't bother copying the Cygwin setup program
## [why]

Unfortunately, the canonical Cygwin setup application forks another process to complete
the package installation process and quickly terminates (even when used with the `--wait`
option). GitHub Actions (GHA) only follows the that initial process and so shuts down
the job step right after the initial process returns, long before the actual installation
is complete. So, the setup program can't be used to install additional packages.

Fortunately, for the Windows MinGW toolsets, the 'depexts-cygwinports' opam package
(https://github.com/fdopen/depext-cygwinports), which is installed here for all Windows
MinGW toolsets, brings with it two utilities (`cygwin-dl` and `cygwin-install`) which
can be used as effective replacements for the setup program.